+++ dwertheimer.EventAutomations/src/index.js
@@ -1,6 +1,6 @@
 // @flow
 import pluginJson from '../plugin.json'
-import { migrateConfiguration, updateSettingData } from '../../helpers/NPConfiguration'
+import { migrateConfiguration, updateSettingData, pluginUpdated } from '../../helpers/NPConfiguration'
 
 export {
   insertTodosAsTimeblocks,
@@ -11,6 +11,7 @@ export {
   removeTimeBlocks,
   removePreviousSyncedCopies,
   removePreviousTimeBlocks,
+  updateDatePlusTags,
 } from './NPTimeblocking'
 
 const PLUGIN_ID = 'autoTimeBlocking' // the key that's used in _configuration note
@@ -32,5 +33,12 @@ export async function onUpdateOrInstall(config: any = { silent: false }): Promis
 }
 
 export async function onSettingsUpdated() {
-  console.log(`${PLUGIN_ID}: onSettingsUpdated ran - nothing to do`)
+  // console.log(`${PLUGIN_ID}: onSettingsUpdated ran - nothing to do`)
+}
+
+export function init(): void {
+  // this runs every time the plugin starts up (any command in this plugin is run)
+  DataStore.installOrUpdatePluginsByID([pluginJson['plugin.id']], true, false, false).then((r) =>
+    pluginUpdated(pluginJson, r),
+  )
 }